requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-results==2.5.1
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.1.0
python-telegram-bot==20.7
python-dotenv==1.0.0
dj-database-url==2.0.0
gunicorn==21.2.0
whitenoise==6.5.0
drf-yasg==1.21.7
django-filter==23.5
django-allauth==0.58.2
django-extensions==3.2.3